[WIP] added onDelete firebase triggers for models that will handle association and file clean-up
