+++ functions/admin.js
@@ -1,0 +1,10 @@
+const functions = require('firebase-functions')
+const admin = require('firebase-admin')
+
+admin.initializeApp()
+const db = admin.firestore()
+
+module.exports = {
+  admin,
+  db
+}
