+++ src/store/teams.ts
@@ -63,7 +63,6 @@ export const useTeamsStore = defineStore('teams', {
     async deleteTeam(team:Team):Promise<void> {
       try {
         await firestore.deleteTeam(team)
-        await this.removeTeamAvatar(team)
       }
       catch(error) {
         let message = (error instanceof Error) ? error.message : String(error)
